restructured shema definations
