+++ collector/graphql/mutations/index.js
@@ -1,0 +1,5 @@
+const UserMutation = require('./user');
+
+module.exports = [
+  UserMutation
+]
