+++ collector/graphql/queries/index.js
@@ -1,0 +1,5 @@
+const UserQuery = require('./user');
+
+module.exports = [
+  UserQuery
+]
