+++ collector/graphql/types/file.js
@@ -1,0 +1,13 @@
+const { gql } = require('apollo-server-express');
+
+const FileType =  gql`
+  scalar Upload
+  
+  type File {
+    filename: String!
+    mimetype: String!
+    encoding: String!
+  }
+`;
+
+module.exports = FileType;
