+++ collector/graphql/types/index.js
@@ -1,0 +1,5 @@
+const UserType = require('./user');
+
+module.exports = [
+  UserType
+]
